3-fastapi/pyproject.toml
[tool.poetry]
name = "aws-blue-green-sandbox"
version = "0.1.0"
description = "Predictor on AWS Apprunner"
authors = ["chck <shimekiri.today@gmail.com>"]
readme = "README.md"

[tool.taskipy.tasks]
dev = {cmd = "uvicorn main:app --reload", help = "run application for development"}
lint = {cmd = "pysen run lint", help = "execute linter"}
fmt = {cmd = "pysen run format", help = "format code via linter"}
test = {cmd = "task fmt && pytest -s", help = "run all tests"}

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
gunicorn = "^20.1.0"
pandas = "^1.4.2"
redis = "^4.3.1"
matplotlib = "^3.5.2"
fasttext-wheel = "^0.9.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
taskipy = "^1.10.1"
pysen = {extras = ["lint"], version = "^0.10.4"}
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.1"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 120
py_version = "py39"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pytest.ini_options]
norecursedirs = [".git", ".tox", "venv*", "requirements*"]
filterwarnings = "error"
addopts = """
--cov-report term-missing \
--cov main.py -ra \
-n auto"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
